- Adding Home Page - Refactoring Login, MasterDetail  and Menu pages. - Adding Navigation Page
