+++ ProyectoXamarin/ProyectoXamarin.Android/ProyectoXamarin.Android.csproj.bak
@@ -8,7 +8,7 @@
     <OutputType>Library</OutputType>
     <RootNamespace>ProyectoXamarin.Droid</RootNamespace>
     <AssemblyName>ProyectoXamarin.Android</AssemblyName>
-    <TargetFrameworkVersion>v8.0</TargetFrameworkVersion>
+    <TargetFrameworkVersion>v7.1</TargetFrameworkVersion>
     <AndroidApplication>True</AndroidApplication>
     <AndroidResgenFile>Resources\Resource.designer.cs</AndroidResgenFile>
     <AndroidResgenClass>Resource</AndroidResgenClass>
