+++ ProyectoXamarin/ProyectoXamarin/Views/HomePage.xaml
@@ -1,0 +1,27 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             x:Class="ProyectoXamarin.Views.HomePage">
+    <ContentPage.Content>
+        <StackLayout VerticalOptions="Center" HorizontalOptions="Center" Margin="10">
+            <Grid>
+                <Grid.RowDefinitions>
+                    <RowDefinition Height="80"/>
+                    <RowDefinition Height="80"/>
+                    <RowDefinition Height="80"/>
+                    <RowDefinition Height="80"/>
+                </Grid.RowDefinitions>
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition/>
+                    <ColumnDefinition/>
+                </Grid.ColumnDefinitions>
+                <Label Text="Bienvenido - Aula Virtual" VerticalOptions="Center" HorizontalOptions="Center" Grid.Row="0" Grid.ColumnSpan="2" FontSize="Large"/>
+                <Button Grid.Row="1" Grid.Column="0" Text="Iniciar Sesión" Grid.ColumnSpan="2" />
+                <Button Grid.Row="2" Grid.Column="0" Text="Profesores"/>
+                <Button Grid.Row="2" Grid.Column="1" Text="Secciones"/>
+                <Button Grid.Row="3" Grid.Column="0" Text="Clases"/>
+                <Button Grid.Row="3" Grid.Column="1" Text="Ayuda"/>
+            </Grid>
+        </StackLayout>
+    </ContentPage.Content>
+</ContentPage>
