+++ ProyectoXamarin/ProyectoXamarin/Views/MasterDetailPage.xaml
@@ -2,9 +2,8 @@
 <MasterDetailPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              x:Class="ProyectoXamarin.Views.MasterDetailPage"
-              xmlns:local="clr-namespace:ProyectoXamarin.Views"    >
-    
+              xmlns:local="clr-namespace:ProyectoXamarin.Views">
     <MasterDetailPage.Detail>
-        <local:RegistrationPage/>
+        <local:HomePage/>
     </MasterDetailPage.Detail>
 </MasterDetailPage>
