+++ ProyectoXamarin/ProyectoXamarin/Views/MasterDetailPage.xaml.cs
@@ -1,48 +1,49 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-using Xamarin.Forms;
+﻿using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
 
 namespace ProyectoXamarin.Views
 {
-	[XamlCompilation(XamlCompilationOptions.Compile)]
-	public partial class MasterDetailPage
-	{
-		public MasterDetailPage ()
-		{
-			InitializeComponent ();
+    [XamlCompilation(XamlCompilationOptions.Compile)]
+    public partial class MasterDetailPage
+    {
+        public MasterDetailPage()
+        {
+            InitializeComponent();
 
-            var paginaMenu = new MenuPage();
+            var menuPage = new MenuPage();
 
-            paginaMenu.Menu.Add(new MenuItem()
+            menuPage.Menu.Add(new MenuItem()
             {
                 Icon = "",
-                Title = "Log in"
+                Title = "Log in",
+                Command = new Command(() => GoToPage(new LoginPage()))
             });
 
-            paginaMenu.Menu.Add(new MenuItem()
+            menuPage.Menu.Add(new MenuItem()
             {
                 Icon = "",
                 Title = "Profesores"
             });
 
-            paginaMenu.Menu.Add(new MenuItem()
+            menuPage.Menu.Add(new MenuItem()
             {
                 Icon = "",
                 Title = "Secciones"
             });
 
-            paginaMenu.Menu.Add(new MenuItem()
+            menuPage.Menu.Add(new MenuItem()
             {
                 Icon = "",
                 Title = "Clases"
             });
 
-            Master = paginaMenu;
+            Master = menuPage;
         }
-	}
+
+        private void GoToPage(Page page)
+        {
+            Detail = page;
+            IsPresented = false;
+        }
+    }
 }
