+++ ProyectoXamarin/ProyectoXamarin/Views/MenuPage.xaml
@@ -1,32 +1,28 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-              x:Class="ProyectoXamarin.Views.MenuPage">
+             x:Class="ProyectoXamarin.Views.MenuPage"
              Title="Menu">
-    <ContentPage.Content>
-        <ListView x:Name="MenuList"
-                  ItemsSource="{Binding Menu}"
-                  HasUnevenRows="True">
-            <ListView.ItemTemplate>
-                <DataTemplate>
-                    <ViewCell>
-                        <Grid>
-                            <Grid.ColumnDefinitions>
-                                <ColumnDefinition Width="Auto"></ColumnDefinition>
-                                <ColumnDefinition Width="*"></ColumnDefinition>
-                            </Grid.ColumnDefinitions>
-                            <Image 
-                                  WidthRequest="50"
-                                  HeightRequest="50"
-                                  Source="{Binding Icon}"></Image>
-                            <Label 
-                                  Grid.Column="1"
-                                  VerticalOptions="Center"
-                                  Text="{Binding Title}"></Label>
-                        </Grid>
-                    </ViewCell>
-                </DataTemplate>
-            </ListView.ItemTemplate>
-        </ListView>
-    </ContentPage.Content>
+    <ListView   x:Name="MenuList"
+                ItemsSource="{Binding Menu}"
+                HasUnevenRows="True">
+        <ListView.ItemTemplate>
+            <DataTemplate>
+                <ViewCell>
+                    <Grid>
+                        <Grid.ColumnDefinitions>
+                            <ColumnDefinition Width="Auto"></ColumnDefinition>
+                            <ColumnDefinition Width="*"></ColumnDefinition>
+                        </Grid.ColumnDefinitions>
+                        <Image  WidthRequest="50"
+                                HeightRequest="50"
+                                Source="{Binding Icon}"></Image>
+                        <Label  Grid.Column="1"
+                                VerticalOptions="Center"
+                                Text="{Binding Title}"></Label>
+                    </Grid>
+                </ViewCell>
+            </DataTemplate>
+        </ListView.ItemTemplate>
+    </ListView>
 </ContentPage>
